1.0

+++ src/services/contacts.js
@@ -8,11 +8,12 @@ export const getAllContacts = async ({
   sortOrder = SORT_ORDER.ASC,
   sortBy = '_id',
   filter = {},
+  userId,
 }) => {
   const limit = perPage;
   const skip = (page - 1) * perPage;
 
-  const contactsQuery = ContactsCollection.find();
+  const contactsQuery = ContactsCollection.find({ userId });
 
   if (filter.contactType) {
     contactsQuery.where('contactType').equals(filter.contactType);
@@ -39,28 +40,28 @@ export const getAllContacts = async ({
   };
 };
 
-export const getContactById = async (contactId) => {
-  const contact = await ContactsCollection.findById(contactId);
+export const getContactById = async (contactId, userId) => {
+  const contact = await ContactsCollection.findOne({ userId, _id: contactId });
   return contact;
 };
 
-export const createContact = async (payload) => {
-  const contact = ContactsCollection.create(payload);
+export const createContact = async (payload, userId) => {
+  const contact = ContactsCollection.create({ ...payload, userId });
   return contact;
 };
 
-export const deleteContact = async (contactId) => {
-  const contact = ContactsCollection.findOneAndDelete({ _id: contactId });
+export const deleteContact = async (contactId, userId) => {
+  const contact = ContactsCollection.findOneAndDelete({ _id: contactId, userId });
   return contact;
 };
 
-export const updateContact = async (contactId, payload) => {
-  const result = await ContactsCollection.findOneAndUpdate({ _id: contactId }, payload, {
+export const updateContact = async (contactId, userId, payload) => {
+  if (!userId) return;
+
+  const result = await ContactsCollection.findOneAndUpdate({ userId, _id: contactId }, payload, {
     new: true,
     includeResultMetadata: true,
   });
-
-  console.log(result);
 
   if (!result || !result.value) return null;
